Construct

Topics

data Topic

Topics in a mindmap

topic :: String -> Topic

Relations

data Relation

Directed relations between topics with descriptions

Relation Operators

(|->) :: (String, String) -> String -> Relation

Arrow operator for creating relations

(-|) :: String -> String -> (String, String)

auxiliary operator for use with the arrow operator

Mindmap

data Mindmap

A mindmap in its most basic form consists of a collection of topics and their relations. These can be further augmented to include groupings, colors, and other layout or appearance features.